Do naive matrix multiplication instead of calling Eigen functions
In the csm_eigen branch of csm, which is what we use, the gsl calls wrap
Eigen, so `m_mult`, rather than calling gsl_blas_dgemm, calls `*` on the
internally-Eigen matrices.

When we compile with clang, issues arise deep inside the Eigen matrix
multiplication code. The members holding the number of rows and columns
get corrupted (or are not initialized) and end up holding random values.
Depending on luck, this will cause the multiply method to segfault,
spend forever multiplying impossibly large matrices, or seemingly work.

The problem is likely a memory error somewhere, but I haven't found it.
Not calling the Eigen multiply circumvents the issue and lets us compile
& run the code though.

For small matrices, the naive code will be faster, while for large
matrices, Eigen's code is likely(?) faster. It seems that we do (on the
order of) 40 2x2 multiplies followed by two 12x12 multiplies per
scanmatch.